updated png to webp for few images + css file
