+++ src/components/AboutBlurb/AboutBlurb.css
@@ -15,8 +15,7 @@
 }
 
 #cardIcon {
-  width: 4.5vw;
-  height: 4.5vw;
+  width: 4.5vh;
 }
 
 #stackityStack {
