+++ src/components/DataVisual/DataVisual.css
@@ -8,8 +8,8 @@
 }
 
 #cardIcon {
-  width: 4.5vh;
-  height: 4.5vh;
+  font-size: 3rem;
+  width: 7.85vw;
   margin-left: auto;
   margin-right: auto;
 }
@@ -18,18 +18,19 @@
   text-align: center;
   margin-left: auto;
   margin-right: auto;
-  font-size: 3.5vh;
+  font-size: 2.3rem;
 }
 
 #cardText {
   margin-left: auto;
   margin-right: auto;
   text-align: center;
-  font-size: 15px;
+  font-size: 1rem;
 }
 
 #stackityStack {
-  margin: 0.3em 1.5em;
+  margin-top: -1.5em;
   justify-content: space-around;
+  flex-direction: row;
   align-items: flex-start;
 }
